requirements.txt
dotenv
langchain_google_genai
langchain_core
langchain_community
PyPDFLoader
rapidocr-onnxruntime
langchain-chroma
duckduckgo-search